refactor: split openapi export command
