+++ backend/commands/__init__.py
@@ -1,0 +1,1 @@
+"""バックエンド運用コマンドをまとめる。"""
