+++ backend/presentation/openapi.py
@@ -1,4 +1,3 @@
-import argparse
 import json
 from pathlib import Path
 from typing import Any
@@ -21,22 +20,3 @@ def write_openapi_schema(output_path: str | Path) -> Path:
         encoding="utf-8",
     )
     return path
-
-
-def parse_args() -> argparse.Namespace:
-    parser = argparse.ArgumentParser(description="FastAPI の OpenAPI spec を JSON として出力します。")
-    parser.add_argument(
-        "--output",
-        default="backend/openapi.json",
-        help="出力先の JSON ファイルパス",
-    )
-    return parser.parse_args()
-
-
-def main() -> None:
-    args = parse_args()
-    write_openapi_schema(args.output)
-
-
-if __name__ == "__main__":
-    main()
